refactor(Effects): move OnRunEffects to lifecycle_hooks
