+++ modules/effects/src/index.ts
@@ -7,8 +7,7 @@ export { mergeEffects } from './effects_resolver';
 export { Actions, ofType } from './actions';
 export { EffectsModule } from './effects_module';
 export { EffectSources } from './effect_sources';
-export { OnRunEffects } from './on_run_effects';
 export { EffectNotification } from './effect_notification';
 export { ROOT_EFFECTS_INIT } from './effects_root_module';
 export { UPDATE_EFFECTS, UpdateEffects } from './effects_feature_module';
-export { OnIdentifyEffects } from './lifecycle_hooks';
+export { OnIdentifyEffects, OnRunEffects } from './lifecycle_hooks';
