+++ modules/effects/src/lifecycle_hooks.ts
@@ -1,3 +1,6 @@
+import { Observable } from 'rxjs';
+import { EffectNotification } from '.';
+
 /**
  * @description
  * Interface to set an identifier for effect instances.
@@ -35,3 +38,44 @@ export interface OnIdentifyEffects {
 
 export const onIdentifyEffectsKey: keyof OnIdentifyEffects =
   'ngrxOnIdentifyEffects';
+
+export type onRunEffectsFn = (
+  resolvedEffects$: Observable<EffectNotification>
+) => Observable<EffectNotification>;
+
+/**
+ * @description
+ * Interface to control the lifecycle of effects.
+ *
+ * By default, effects are merged and subscribed to the store. Implement the OnRunEffects interface to control the lifecycle of the resolved effects.
+ *
+ * @usageNotes
+ *
+ * ### Implement the OnRunEffects interface on an Effects class
+ *
+ * ```ts
+ * export class UserEffects implements OnRunEffects {
+ *   constructor(private actions$: Actions) {}
+ *
+ *   ngrxOnRunEffects(resolvedEffects$: Observable<EffectNotification>) {
+ *     return this.actions$.pipe(
+ *       ofType('LOGGED_IN'),
+ *       exhaustMap(() =>
+ *         resolvedEffects$.pipe(
+ *           takeUntil(this.actions$.pipe(ofType('LOGGED_OUT')))
+ *         )
+ *       )
+ *     );
+ *   }
+ * }
+ * ```
+ */
+export interface OnRunEffects {
+  /**
+   * @description
+   * Method to control the lifecycle of effects.
+   */
+  ngrxOnRunEffects: onRunEffectsFn;
+}
+
+export const onRunEffectsKey: keyof OnRunEffects = 'ngrxOnRunEffects';
